+++ includes/entidades/obstaculos/Arbusto.hpp
@@ -9,10 +9,10 @@ namespace Entidades
         class Arbusto : public Obstaculo
         {
         private:
-            int largura;
+            float largura;
 
         public:
-            Arbusto(int posX = 0, int posY = 0, int tamanhoX = 0, int tamanhoY = 0);
+            Arbusto(float posX = 0.0f, float posY = 0.0f, float tamanhoX = 0.0f, float tamanhoY = 0.0f);
             ~Arbusto();
             void salvar();
             void executar(Entidades::Personagens::Jogador *p);
