+++ includes/entidades/personagens/ET_Varginha.hpp
@@ -14,7 +14,7 @@ namespace Entidades
             int multiplicador_laser;
 
         public:
-            ET_Varginha(int posX = 0, int posY = 0, int n = 3, int maldade = 0, int ml = 0);
+            ET_Varginha(float posX = 0.0f, float posY = 0.0f, int n = 3, int maldade = 0, int ml = 0);
             ~ET_Varginha();
 
             // void salvarDataBuffer();
